Produced the data in https://github.com/kokkos/kokkos/issues/8506

```
--------------------------------------------------------------------------------------------------------
Benchmark                                                              Time             CPU   Iterations
--------------------------------------------------------------------------------------------------------
GemvDefault<Kokkos::LayoutLeft>/M:4000/N:10000/manual_time          2629 us         2610 us          266
GemvDefault<Kokkos::LayoutLeft>/M:400000/N:100/manual_time          59.5 us         58.1 us        11882
GemvDefault<Kokkos::LayoutRight>/M:4000/N:10000/manual_time         3413 us         3383 us          205
GemvDefault<Kokkos::LayoutRight>/M:400000/N:100/manual_time          396 us          391 us         1911
Gemv<64, Kokkos::LayoutLeft>/M:4000/N:10000/manual_time             2345 us         2321 us          298
Gemv<64, Kokkos::LayoutLeft>/M:400000/N:100/manual_time             63.1 us         61.2 us        11185
Gemv<64, Kokkos::LayoutRight>/M:4000/N:10000/manual_time             640 us          632 us         1093
Gemv<64, Kokkos::LayoutRight>/M:400000/N:100/manual_time             364 us          361 us         1878
Gemv<256, Kokkos::LayoutLeft>/M:4000/N:10000/manual_time            2513 us         2493 us          279
Gemv<256, Kokkos::LayoutLeft>/M:400000/N:100/manual_time            50.7 us         50.3 us        13777
Gemv<256, Kokkos::LayoutRight>/M:4000/N:10000/manual_time            856 us          848 us          820
Gemv<256, Kokkos::LayoutRight>/M:400000/N:100/manual_time            354 us          350 us         1992
Gemv<1024, Kokkos::LayoutLeft>/M:4000/N:10000/manual_time           2616 us         2592 us          267
Gemv<1024, Kokkos::LayoutLeft>/M:400000/N:100/manual_time           51.0 us         50.7 us        13655
Gemv<1024, Kokkos::LayoutRight>/M:4000/N:10000/manual_time          3402 us         3379 us          206
Gemv<1024, Kokkos::LayoutRight>/M:400000/N:100/manual_time           384 us          382 us         1974
GemvHip<64, Kokkos::LayoutLeft>/M:4000/N:10000/manual_time          2312 us         2296 us          302
GemvHip<64, Kokkos::LayoutLeft>/M:400000/N:100/manual_time          51.2 us         50.9 us        13678
GemvHip<64, Kokkos::LayoutRight>/M:4000/N:10000/manual_time          598 us          593 us         1165
GemvHip<64, Kokkos::LayoutRight>/M:400000/N:100/manual_time          535 us          532 us         1306
GemvHip<256, Kokkos::LayoutLeft>/M:4000/N:10000/manual_time         2289 us         2272 us          306
GemvHip<256, Kokkos::LayoutLeft>/M:400000/N:100/manual_time         50.6 us         50.3 us        13806
GemvHip<256, Kokkos::LayoutRight>/M:4000/N:10000/manual_time         593 us          589 us         1175
GemvHip<256, Kokkos::LayoutRight>/M:400000/N:100/manual_time         543 us          539 us         1290
GemvHip<1024, Kokkos::LayoutLeft>/M:4000/N:10000/manual_time        2524 us         2506 us          277
GemvHip<1024, Kokkos::LayoutLeft>/M:400000/N:100/manual_time        50.4 us         50.1 us        13815
GemvHip<1024, Kokkos::LayoutRight>/M:4000/N:10000/manual_time       2737 us         2716 us          256
GemvHip<1024, Kokkos::LayoutRight>/M:400000/N:100/manual_time        483 us          479 us         1439
```