@cwpearson 2025-10-01
I don't see any interaction with `ChunkSize` / `chunks_size` in e.g. https://github.com/kokkos/kokkos/blob/b3ec45c33bc8151b92d63a764f01281911c83cf9/core/src/HIP/Kokkos_HIP_BlockSize_Deduction.hpp#L23-L35 so I think for HIP we only have the LaunchBounds mechanism.