@Rombur 2025-10-01
`ChunkSize` controls the block size **only** for SYCL not for the other backends.